Adding error handling for cases when the briefcase contains deleted changesets

+++ core/backend/src/ConcurrencyControl.ts
@@ -1578,6 +1578,9 @@ export namespace ConcurrencyControl { // eslint-disable-line no-redeclare
     }
 
     public close(saveChanges: boolean) {
+      if (!this._locksFileName || !IModelJsFs.existsSync(this._locksFileName))
+        return;
+
       if (saveChanges)
         this._db.saveChanges();
       else
@@ -1601,6 +1604,10 @@ export namespace ConcurrencyControl { // eslint-disable-line no-redeclare
         if (DbResult.BE_SQLITE_DONE !== rc)
           throw new IModelError(rc, "", Logger.logError, loggerCategory, () => sql);
       });
+    }
+
+    private doesCacheFileExist(): boolean {
+      return this._locksFileName !== undefined && IModelJsFs.existsSync(this._locksFileName);
     }
 
     public open(): boolean {
@@ -1640,6 +1647,9 @@ export namespace ConcurrencyControl { // eslint-disable-line no-redeclare
     }
 
     public deleteFile() {
+      if (!this.doesCacheFileExist())
+        return;
+
       if (this.isOpen)
         this.close(false);
 
@@ -1717,6 +1727,9 @@ export namespace ConcurrencyControl { // eslint-disable-line no-redeclare
     }
 
     public deleteLocksForTxn(txnId: string) {
+      if (!this.doesCacheFileExist())
+        return;
+
       this.mustBeOpenAndWriteable();
       this._db.withPreparedSqliteStatement("delete from heldLocks where txnId=?", (stmt) => {
         stmt.bindValue(1, txnId);
@@ -1749,7 +1762,6 @@ export namespace ConcurrencyControl { // eslint-disable-line no-redeclare
       this.saveChanges();
     }
   }
-
 }
 
 /**
